requirements.txt
streamlit==1.38.0
pymongo==4.8.0 
aiohttp==3.10.5 
python-dotenv==1.0.1